pop PronounSet and PronounTrie into lib.rs
This just gives each component a bit more breathing room.
